barcode_project/requirements.txt
python-barcode
reportlab
Pillow
